Refactor Nostr profile handling: remove unused cache follows hook, implement batch profile fetching, and update components to use new hook
