+++ hooks/nostrProfile/index.ts
@@ -2,5 +2,4 @@ export * from "./useNostrProfile";
 export * from "./useSaveNostrProfile";
 export * from "./useAddFollow";
 export * from "./useRemoveFollow";
-export * from "./useCacheFollows";
 export * from "./useIsPubkeyFollowed";
