+++ hooks/nostrProfile/useNostrProfile.ts
@@ -4,8 +4,10 @@ import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useCallback } from "react";
 import { Event } from "nostr-tools";
 
-const STALE_TIME = 24 * 60 * 60 * 1000; // 24 hours in milliseconds
-type NostrUserProfileWithTimestamp = NostrUserProfile & { created_at: number };
+export const STALE_TIME = 24 * 60 * 60 * 1000; // 24 hours in milliseconds
+export type NostrUserProfileWithTimestamp = NostrUserProfile & {
+  created_at: number;
+};
 
 export function useNostrProfile(pubkey?: string | null, relays?: string[]) {
   const queryClient = useQueryClient();
@@ -65,12 +67,12 @@ export function useNostrProfile(pubkey?: string | null, relays?: string[]) {
 
             return getLatestEvent(filter, relayList);
           },
-          structuralSharing: (prev: any, next: any) => {
-            if ("created_at" in next && "created_at" in prev) {
-              return next.created_at > prev.created_at ? next : prev;
-            }
-            return next;
-          },
+          // structuralSharing: (prev: any, next: any) => {
+          //   if ("created_at" in next && "created_at" in prev) {
+          //     return next.created_at > prev.created_at ? next : prev;
+          //   }
+          //   return next;
+          // },
           staleTime: STALE_TIME,
           gcTime: Infinity,
         });
@@ -81,61 +83,9 @@ export function useNostrProfile(pubkey?: string | null, relays?: string[]) {
     [queryClient, getLatestEvent],
   );
 
-  const batchGetProfileMetadata = useCallback(
-    async (pubkeys: string[], relayList?: string[]) => {
-      // Skip empty requests
-      if (!pubkeys.length)
-        return new Map<string, NostrUserProfileWithTimestamp>();
-
-      // Check cache first and collect missing pubkeys
-      const profiles = new Map<string, NostrUserProfileWithTimestamp>();
-      const missingPubkeys: string[] = [];
-
-      pubkeys.forEach((pubkey) => {
-        const queryKey = nostrQueryKeys.profile(pubkey);
-        const cachedData = queryClient.getQueryData(queryKey);
-        const queryState = queryClient.getQueryState(queryKey);
-        const dataAge = queryState?.dataUpdatedAt
-          ? Date.now() - queryState.dataUpdatedAt
-          : Infinity;
-
-        if (cachedData && dataAge <= STALE_TIME) {
-          profiles.set(pubkey, cachedData as NostrUserProfileWithTimestamp);
-        } else {
-          missingPubkeys.push(pubkey);
-        }
-      });
-
-      // If we have missing profiles, fetch them in a single query
-      if (missingPubkeys.length > 0) {
-        const filter = {
-          kinds: [0],
-          authors: missingPubkeys,
-        };
-
-        const { querySync } = useNostrEvents();
-        const events = await querySync(filter, relayList);
-
-        // Process and cache the results
-        events.forEach((event) => {
-          const profile = decodeProfileMetadata(event);
-          if (profile) {
-            const queryKey = nostrQueryKeys.profile(event.pubkey);
-            queryClient.setQueryData(queryKey, profile);
-            profiles.set(event.pubkey, profile);
-          }
-        });
-      }
-
-      return profiles;
-    },
-    [queryClient],
-  );
-
   return {
     ...queryData,
     getProfileMetadata,
-    batchGetProfileMetadata,
     decodeProfileMetadata,
   };
 }
